Google analytics och lang="en"
