+++ pages/_document.js
@@ -1,0 +1,38 @@
+import Document, { Html, Head, Main, NextScript } from 'next/document'
+
+class MyDocument extends Document {
+    static async getInitialProps(ctx) {
+        const initialProps = await Document.getInitialProps(ctx)
+        return { ...initialProps }
+    }
+
+    render() {
+        return (
+            <Html lang="en">
+                <Head>
+                    <script async src="https://www.googletagmanager.com/gtag/js?id=G-MT5NWW6HW9"></script>
+                    <script
+                        dangerouslySetInnerHTML={{
+                            __html: `
+                                window.dataLayer = window.dataLayer || [];
+                                function gtag(){dataLayer.push(arguments);}
+                                gtag('js', new Date());
+                                gtag('config', 'G-MT5NWW6HW9', {
+                                    'linker': {
+                                    'domains': ['engineeringguidance.com', 'engui.xyz']
+                                    }
+                                );
+                            `
+                        }}
+                    />
+                </Head>
+                <body>
+                    <Main />
+                    <NextScript />
+                </body>
+            </Html>
+        )
+    }
+}
+
+export default MyDocument
